+++ Sources/SentryCrash/Recording/SentryCrashReport.c
@@ -1185,6 +1185,10 @@ static void
 binaryImagesIteratorCallback(SentryCrashBinaryImage *image, void *context)
 {
     SentryCrashReportWriter *writer = (SentryCrashReportWriter *)context;
+    // We can only retrieve the crash info after a crash occurred. So we need to
+    // fetch it when writing the crash report.
+    // Swift puts its fatalErrors into the crash info message.
+    sentrycrashdl_getCrashInfo(image->address, image);
     writeBinaryImage(writer, NULL, image);
 }